requirements.txt
python-telegram-bot==21.0.1
supabase==2.28.3
anthropic>=0.40.0
APScheduler==3.10.4
python-dotenv==1.0.0
feedparser>=6.0.0
httpx>=0.27.0
